// fe/src/lib/api/generated/model/organizationMembershipPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { OrganizationScopeValue } from './organizationScopeValue';

export interface OrganizationMembershipPayload {
  organization_name?: string;
  position_title?: string;
  scope: OrganizationScopeValue;
  /** @minimum 0 */
  sort_order: number;
}