// fe/src/lib/api/generated/model/transportationEntryPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { TransportationFrequencyCategoryValue } from './transportationFrequencyCategoryValue';
import type { TransportationModeValue } from './transportationModeValue';

export interface TransportationEntryPayload {
  fare?: string | null;
  frequency?: string;
  frequency_category?: TransportationFrequencyCategoryValue | null;
  mode: TransportationModeValue;
}